nvim: add custom color highlighter
First iteration of my custom theme. The idea is to have a general highlighter and define the base 16 color hash.
In the future, the idea is to have one file to hold the colors and generate the input for nvim and kitty, to be able to change both at the same time.